% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack7.R
\name{cheat}
\alias{cheat}
\title{Allows the user of this package to cheat on assignment 3.2.1 for question 1,2,6,16}
\usage{
cheat(number)
}
\arguments{
\item{number}{numeric variable}
}
\description{
Allows the user of this package to cheat on assignment 3.2.1 for question 1,2,6,16
}
